+++ app/Http/Controllers/DepartmentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Department;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use App\Models\Role;
 use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -12,11 +13,11 @@ class DepartmentController extends Controller
 {
     public function index()
     {
-        // abort_if(
-        //     Gate::denies('position_access'),
-        //     Response::HTTP_FORBIDDEN,
-        //     '403 Forbidden'
-        // );
+        abort_if(
+            Gate::denies('department_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
 
         $departments = Department::all();
         $canCreate = Gate::allows('position_create');
@@ -26,24 +27,37 @@ class DepartmentController extends Controller
 
     public function store(Request $request)
     {
-        // abort_if(
-        //     Gate::denies('position_create'),
-        //     Response::HTTP_FORBIDDEN,
-        //     '403 Forbidden'
-        // );
+        abort_if(
+            Gate::denies('department_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
 
         $request->validate([
             'title' => 'required'
         ]);
 
-        Department::create([
-            'title' => $request->title
-        ]);
+        if($request->dept_id == ''){
+            Department::create([
+                'title' => $request->title
+            ]);
+        }else{
+            Department::where('id', $request->dept_id)->update([
+                'title' => $request->title
+            ]);
+        }
+
         return redirect()->route('department');
     }
 
     public function destroy(Request $request, $id)
     {
+        abort_if(
+            Gate::denies('department_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+        
         Department::find($id)->delete();
 
         return redirect()->route('department');
