+++ app/Http/Controllers/EmployeeStatusController.php
@@ -12,11 +12,11 @@ class EmployeeStatusController extends Controller
 {
     public function index()
     {
-        // abort_if(
-        //     Gate::denies('position_access'),
-        //     Response::HTTP_FORBIDDEN,
-        //     '403 Forbidden'
-        // );
+        abort_if(
+            Gate::denies('employee_status_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
 
         $statuses = EmployeeStatus::all();
         $canCreate = Gate::allows('position_create');
@@ -27,7 +27,7 @@ class EmployeeStatusController extends Controller
     public function create()
     {
         abort_if(
-            Gate::denies('position_create'),
+            Gate::denies('employee_status_access'),
             Response::HTTP_FORBIDDEN,
             '403 Forbidden'
         );
@@ -37,11 +37,11 @@ class EmployeeStatusController extends Controller
 
     public function store(Request $request)
     {
-        // abort_if(
-        //     Gate::denies('position_create'),
-        //     Response::HTTP_FORBIDDEN,
-        //     '403 Forbidden'
-        // );
+        abort_if(
+            Gate::denies('employee_status_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
 
         $request->validate([
             'title' => 'required'
@@ -55,6 +55,12 @@ class EmployeeStatusController extends Controller
 
     public function destroy(Request $request, $id)
     {
+        abort_if(
+            Gate::denies('employee_status_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+        
         EmployeeStatus::find($id)->delete();
 
         return redirect()->route('settings.emp.statuses');
