+++ app/Http/Controllers/PositionController.php
@@ -12,11 +12,11 @@ class PositionController extends Controller
 {
     public function index()
     {
-        // abort_if(
-        //     Gate::denies('position_access'),
-        //     Response::HTTP_FORBIDDEN,
-        //     '403 Forbidden'
-        // );
+        abort_if(
+            Gate::denies('position_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
 
         $positions = Position::all();
         $canCreate = Gate::allows('position_create');
@@ -26,24 +26,37 @@ class PositionController extends Controller
 
     public function store(Request $request)
     {
-        // abort_if(
-        //     Gate::denies('position_create'),
-        //     Response::HTTP_FORBIDDEN,
-        //     '403 Forbidden'
-        // );
+        abort_if(
+            Gate::denies('position_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
 
         $request->validate([
             'title' => 'required'
         ]);
 
-        Position::create([
-            'title' => $request->title
-        ]);
+        if($request->position_id == ''){
+            Position::create([
+                'title' => $request->title
+            ]);
+        }else{
+            Position::where('id', $request->position_id)->update([
+                'title' => $request->title
+            ]);
+        }
+
         return redirect()->route('position');
     }
 
     public function destroy(Request $request, $id)
     {
+        abort_if(
+            Gate::denies('position_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+
         Position::find($id)->delete();
 
         return redirect()->route('position');
