+++ app/Http/Controllers/StockTransferController.php
@@ -9,6 +9,9 @@ use App\Models\StockTransferDetail;
 use App\Models\Location;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use App\Models\Role;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 use Illuminate\Support\Facades\DB;
 
@@ -16,6 +19,12 @@ class StockTransferController extends Controller
 {
 	public function index()
     {
+        abort_if(
+            Gate::denies('stock_transfer_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+
         $stock_transfers = StockTransfer::with('location_from','location_to')->get();
         $locations = Location::all();
         return Inertia::render('Stock/Transfer/Index', compact('stock_transfers','locations'));
@@ -23,6 +32,12 @@ class StockTransferController extends Controller
 
     public function store(Request $request)
     {
+        abort_if(
+            Gate::denies('stock_transfer_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+
         $reference = "ST-" . date("Ymdhis");
         $request->validate([
             "from" => ['required'],
@@ -43,6 +58,12 @@ class StockTransferController extends Controller
 
     public function show($stock_transfer_id)
     {
+        abort_if(
+            Gate::denies('stock_transfer_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+
         $stock_transfer = StockTransfer::with('location_from','location_to')->find($stock_transfer_id);
 
         $stocks = StockTransaction::selectRaw("SUM(IF(transaction = 'IN', quantity, -quantity)) as `quantity`,stock_id")
@@ -57,6 +78,12 @@ class StockTransferController extends Controller
     }
 
     public function finish(Request $request){
+        abort_if(
+            Gate::denies('stock_transfer_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+
         StockTransfer::where('id', $request->stock_transfer_id)->update([
             "status" => 1
         ]);
@@ -64,6 +91,12 @@ class StockTransferController extends Controller
     }
 
     public function destroy($id){
+        abort_if(
+            Gate::denies('stock_transfer_access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+        
         StockTransfer::find($id)->delete();
 
         return redirect()->route('stock.transfer');
